# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type PlatformKind = typeof PlatformKind[keyof typeof PlatformKind];


export const PlatformKind = {
  casino: 'casino',
  sports: 'sports',
} as const;

export interface Platform {
  id: string;
  name: string;
  kind: PlatformKind;
  /** @nullable */
  logoUrl: string | null;
  createdAt?: string;
}

export type PlatformInputKind = typeof PlatformInputKind[keyof typeof PlatformInputKind];


export const PlatformInputKind = {
  casino: 'casino',
  sports: 'sports',
} as const;

export interface PlatformInput {
  /** @minLength 1 */
  name: string;
  kind: PlatformInputKind;
  /** @nullable */
  logoUrl?: string | null;
}

export type PlatformUpdateKind = typeof PlatformUpdateKind[keyof typeof PlatformUpdateKind];


export const PlatformUpdateKind = {
  casino: 'casino',
  sports: 'sports',
} as const;

export interface PlatformUpdate {
  /** @minLength 1 */
  name?: string;
  kind?: PlatformUpdateKind;
  /** @nullable */
  logoUrl?: string | null;
}

export type GameKind = typeof GameKind[keyof typeof GameKind];


export const GameKind = {
  casino: 'casino',
  sports: 'sports',
} as const;

export interface Game {
  id: string;
  platformId: string;
  title: string;
  description: string;
  /** @nullable */
  imageUrl: string | null;
  affiliateUrl: string;
  kind: GameKind;
  createdAt?: string;
}

export type GameInputKind = typeof GameInputKind[keyof typeof GameInputKind];


export const GameInputKind = {
  casino: 'casino',
  sports: 'sports',
} as const;

export interface GameInput {
  platformId: string;
  /** @minLength 1 */
  title: string;
  description: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @minLength 1 */
  affiliateUrl: string;
  kind: GameInputKind;
}

export type GameUpdateKind = typeof GameUpdateKind[keyof typeof GameUpdateKind];


export const GameUpdateKind = {
  casino: 'casino',
  sports: 'sports',
} as const;

export interface GameUpdate {
  platformId?: string;
  /** @minLength 1 */
  title?: string;
  description?: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @minLength 1 */
  affiliateUrl?: string;
  kind?: GameUpdateKind;
}

export type ProfileLanguage = typeof ProfileLanguage[keyof typeof ProfileLanguage];


export const ProfileLanguage = {
  RU: 'RU',
  EN: 'EN',
} as const;

export type ProfileTier = typeof ProfileTier[keyof typeof ProfileTier];


export const ProfileTier = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  top: 'top',
} as const;

export interface Profile {
  id: string;
  email: string;
  username: string;
  /** @nullable */
  avatarUrl?: string | null;
  language: ProfileLanguage;
  /** @nullable */
  activePlatformId: string | null;
  referralCode: string;
  referralCount: number;
  balance: number;
  tier: ProfileTier;
  hourlyRate: number;
}

export type ProfileUpdateLanguage = typeof ProfileUpdateLanguage[keyof typeof ProfileUpdateLanguage];


export const ProfileUpdateLanguage = {
  RU: 'RU',
  EN: 'EN',
} as const;

export interface ProfileUpdate {
  /** @minLength 3 */
  username?: string;
  /** @nullable */
  avatarUrl?: string | null;
  language?: ProfileUpdateLanguage;
  /** @nullable */
  activePlatformId?: string | null;
}

export interface UsernameAvailability {
  username: string;
  available: boolean;
}

export interface Dashboard {
  profile: Profile;
  nextTierAt: number;
  referralsToNextTier: number;
  secondsActive: number;
}

export type WithdrawalStatus = typeof WithdrawalStatus[keyof typeof WithdrawalStatus];


export const WithdrawalStatus = {
  pending: 'pending',
  paid: 'paid',
} as const;

export interface Withdrawal {
  id: string;
  username: string;
  amount: number;
  walletAddress: string;
  requestedAt: string;
  status: WithdrawalStatus;
}

export interface WithdrawalInput {
  /** @minimum 0.01 */
  amount: number;
  /** @minLength 20 */
  walletAddress: string;
}

export type WithdrawalStatusUpdateStatus = typeof WithdrawalStatusUpdateStatus[keyof typeof WithdrawalStatusUpdateStatus];


export const WithdrawalStatusUpdateStatus = {
  pending: 'pending',
  paid: 'paid',
} as const;

export interface WithdrawalStatusUpdate {
  status: WithdrawalStatusUpdateStatus;
}

export type ListGamesParams = {
platformId?: string;
kind?: ListGamesKind;
};

export type ListGamesKind = typeof ListGamesKind[keyof typeof ListGamesKind];


export const ListGamesKind = {
  casino: 'casino',
  sports: 'sports',
} as const;

export type CheckUsernameAvailabilityParams = {
username: string;
};