# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: platforms
    description: Casino and bookmaker platforms
  - name: games
    description: Casino games and sports betting events
  - name: profile
    description: Player profile and referral data
  - name: withdrawals
    description: Crypto withdrawal requests
  - name: admin
    description: Restricted administration operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /platforms:
    get:
      operationId: listPlatforms
      tags: [platforms]
      summary: List active platforms
      responses:
        "200":
          description: Platforms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Platform"
    post:
      operationId: createPlatform
      tags: [admin]
      summary: Create a platform
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlatformInput"
      responses:
        "201":
          description: Created platform
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Platform"
  /platforms/{platformId}:
    patch:
      operationId: updatePlatform
      tags: [admin]
      summary: Update a platform
      parameters:
        - $ref: "#/components/parameters/PlatformId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlatformUpdate"
      responses:
        "200":
          description: Updated platform
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Platform"
    delete:
      operationId: deletePlatform
      tags: [admin]
      summary: Delete a platform
      parameters:
        - $ref: "#/components/parameters/PlatformId"
      responses:
        "204":
          description: Deleted
  /games:
    get:
      operationId: listGames
      tags: [games]
      summary: List games and betting events for a platform
      parameters:
        - name: platformId
          in: query
          required: false
          schema:
            type: string
        - name: kind
          in: query
          required: false
          schema:
            type: string
            enum: [casino, sports]
      responses:
        "200":
          description: Games
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Game"
    post:
      operationId: createGame
      tags: [admin]
      summary: Create a game or betting event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameInput"
      responses:
        "201":
          description: Created game
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
  /games/{gameId}:
    patch:
      operationId: updateGame
      tags: [admin]
      summary: Update a game or betting event
      parameters:
        - $ref: "#/components/parameters/GameId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameUpdate"
      responses:
        "200":
          description: Updated game
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
    delete:
      operationId: deleteGame
      tags: [admin]
      summary: Delete a game or betting event
      parameters:
        - $ref: "#/components/parameters/GameId"
      responses:
        "204":
          description: Deleted
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get the current player profile
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
    patch:
      operationId: updateProfile
      tags: [profile]
      summary: Update profile onboarding preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
  /profile/username-availability:
    get:
      operationId: checkUsernameAvailability
      tags: [profile]
      summary: Check whether a username is available
      parameters:
        - name: username
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Availability
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsernameAvailability"
  /profile/dashboard:
    get:
      operationId: getDashboard
      tags: [profile]
      summary: Get dashboard aggregates and reward state
      responses:
        "200":
          description: Dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"
  /withdrawals:
    get:
      operationId: listWithdrawals
      tags: [admin, withdrawals]
      summary: List withdrawal requests
      responses:
        "200":
          description: Withdrawal requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Withdrawal"
    post:
      operationId: createWithdrawal
      tags: [withdrawals]
      summary: Create a pending USDT withdrawal request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalInput"
      responses:
        "201":
          description: Pending withdrawal request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
  /withdrawals/{withdrawalId}:
    patch:
      operationId: updateWithdrawalStatus
      tags: [admin]
      summary: Update payout status
      parameters:
        - $ref: "#/components/parameters/WithdrawalId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalStatusUpdate"
      responses:
        "200":
          description: Updated withdrawal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
components:
  parameters:
    PlatformId:
      name: platformId
      in: path
      required: true
      schema:
        type: string
    GameId:
      name: gameId
      in: path
      required: true
      schema:
        type: string
    WithdrawalId:
      name: withdrawalId
      in: path
      required: true
      schema:
        type: string
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Platform:
      type: object
      required: [id, name, kind, logoUrl]
      properties:
        id: { type: string }
        name: { type: string }
        kind:
          type: string
          enum: [casino, sports]
        logoUrl: { type: ["string", "null"] }
        createdAt: { type: string }
    PlatformInput:
      type: object
      required: [name, kind]
      properties:
        name: { type: string, minLength: 1 }
        kind:
          type: string
          enum: [casino, sports]
        logoUrl: { type: ["string", "null"] }
    PlatformUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        kind:
          type: string
          enum: [casino, sports]
        logoUrl: { type: ["string", "null"] }
    Game:
      type: object
      required: [id, platformId, title, description, imageUrl, affiliateUrl, kind]
      properties:
        id: { type: string }
        platformId: { type: string }
        title: { type: string }
        description: { type: string }
        imageUrl: { type: ["string", "null"] }
        affiliateUrl: { type: string }
        kind:
          type: string
          enum: [casino, sports]
        createdAt: { type: string }
    GameInput:
      type: object
      required: [platformId, title, description, affiliateUrl, kind]
      properties:
        platformId: { type: string }
        title: { type: string, minLength: 1 }
        description: { type: string }
        imageUrl: { type: ["string", "null"] }
        affiliateUrl: { type: string, minLength: 1 }
        kind:
          type: string
          enum: [casino, sports]
    GameUpdate:
      type: object
      properties:
        platformId: { type: string }
        title: { type: string, minLength: 1 }
        description: { type: string }
        imageUrl: { type: ["string", "null"] }
        affiliateUrl: { type: string, minLength: 1 }
        kind:
          type: string
          enum: [casino, sports]
    Profile:
      type: object
      required: [id, email, username, language, activePlatformId, referralCode, referralCount, balance, tier, hourlyRate]
      properties:
        id: { type: string }
        email: { type: string }
        username: { type: string }
        avatarUrl: { type: ["string", "null"] }
        language:
          type: string
          enum: [RU, EN]
        activePlatformId: { type: ["string", "null"] }
        referralCode: { type: string }
        referralCount: { type: integer }
        balance: { type: number }
        tier:
          type: string
          enum: [beginner, intermediate, top]
        hourlyRate: { type: number }
    ProfileUpdate:
      type: object
      properties:
        username: { type: string, minLength: 3 }
        avatarUrl: { type: ["string", "null"] }
        language:
          type: string
          enum: [RU, EN]
        activePlatformId: { type: ["string", "null"] }
    UsernameAvailability:
      type: object
      required: [username, available]
      properties:
        username: { type: string }
        available: { type: boolean }
    Dashboard:
      type: object
      required: [profile, nextTierAt, referralsToNextTier, secondsActive]
      properties:
        profile:
          $ref: "#/components/schemas/Profile"
        nextTierAt: { type: integer }
        referralsToNextTier: { type: integer }
        secondsActive: { type: integer }
    Withdrawal:
      type: object
      required: [id, username, amount, walletAddress, requestedAt, status]
      properties:
        id: { type: string }
        username: { type: string }
        amount: { type: number }
        walletAddress: { type: string }
        requestedAt: { type: string }
        status:
          type: string
          enum: [pending, paid]
    WithdrawalInput:
      type: object
      required: [amount, walletAddress]
      properties:
        amount: { type: number, minimum: 0.01 }
        walletAddress: { type: string, minLength: 20 }
    WithdrawalStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, paid]