# lib/api-zod/src/generated/types/platform.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PlatformKind } from './platformKind';

export interface Platform {
  id: string;
  name: string;
  kind: PlatformKind;
  /** @nullable */
  logoUrl: string | null;
  createdAt?: string;
}