# lib/api-zod/src/generated/types/platformInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PlatformInputKind } from './platformInputKind';

export interface PlatformInput {
  /** @minLength 1 */
  name: string;
  kind: PlatformInputKind;
  /** @nullable */
  logoUrl?: string | null;
}