# lib/api-zod/src/generated/types/platformUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PlatformUpdateKind } from './platformUpdateKind';

export interface PlatformUpdate {
  /** @minLength 1 */
  name?: string;
  kind?: PlatformUpdateKind;
  /** @nullable */
  logoUrl?: string | null;
}