# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProfileLanguage } from './profileLanguage';
import type { ProfileTier } from './profileTier';

export interface Profile {
  id: string;
  email: string;
  username: string;
  /** @nullable */
  avatarUrl?: string | null;
  language: ProfileLanguage;
  /** @nullable */
  activePlatformId: string | null;
  referralCode: string;
  referralCount: number;
  balance: number;
  tier: ProfileTier;
  hourlyRate: number;
}